Move protocols to java.time
Running '=' on the date/time classes didn't work unless a specific
require on cljs.java-time.extend-eq-and-compare was added.

This is a problem when using this library from other libraries,
since it's not obvious and/or easy to forget, and exposes internal
details.

By including it in java.time, all consumers will get correct
equality and hash functionality.

The cljs.java-time.extend-eq-and-compare namespace is emptied and
kept for backward compatibility.